doc: init eg for verifyMessage
