+++ src/util/mattr-client.ts
@@ -1,0 +1,8 @@
+import { MattrViiClient } from '@/client';
+import { ConfigService } from '@nestjs/config';
+
+const config = new ConfigService();
+export const client = new MattrViiClient({
+  tenantUrl: config.get('MATTR_TENANT_URL'),
+  authToken: config.get('MATTR_AUTH_TOKEN'),
+});
